+++ e2e/wallets/intergration-coins/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/integration-coins
 
+## 1.13.20
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.23
+
 ## 1.13.19
 
 ### Patch Changes