+++ packages/pioneer/pioneer-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/pioneer-react
 
+## 0.10.2
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.23
+
 ## 0.10.1
 
 ### Patch Changes